+++ CHANGELOG.md
@@ -263,3 +263,4 @@
 - 2024-01-23: Rate README, chat, campaign, campaigns, +2
 - 2024-01-23: Pagination bettercontact, conf, mutate, follow up, +4
 - 2024-01-25: Cache find email, tz country, daemon, icp, +6
+- 2024-01-25: Cache lead, descend, deal, mod, +6